stj_automation/requirements.txt
selenium==4.27.1
python-dotenv==1.0.1
supabase==2.9.1
requests==2.32.3
webdriver-manager==4.0.2
